… pkg/pb/pipeline/pipeline.pb.go
… pkg/pb/plan/plan.pb.go
… pkg/sql/colexec/rightanti/join_test.go
… pkg/sql/colexec/rightsemi/join_test.go
… pkg/sql/compile/operator.go
… pkg/sql/compile/scopeRemoteRun.go
… pkg/sql/plan/join_order.go
… pkg/sql/plan/query_builder.go
… pkg/sql/plan/stats.go
… pkg/vm/overload.go
… pkg/vm/types.go
… proto/pipeline.proto
… proto/plan.proto