This commit significantly improves both performance and memory for particular TPC-H queries.

Environment: TPC-H SF = 10, AMD Ryzen 5900X + 32GB memory desktop, 1 CN 1 DN

* q4: 6136 ms -> 1244 ms
* q21: 11003 ms -> 3529 ms
* q22: 1685 ms -> 309 ms

And the total memory allocated by hashbuild is reduced from 20266MB to 1189MB for q21.

All but 2 TPC-H SF = 100 queries can be run on my 32GB desktop now. Exceptions are q16, which has `count distinct` aggregate, and q18, which has a middle result of group-by surpassing the 1GB limit of mpool.

## What type of PR is this?

- [ ] API-change
- [ ] BUG
- [ ] Improvement
- [ ] Documentation
- [x] Feature
- [ ] Test and CI
- [ ] Code Refactoring

## Which issue(s) this PR fixes:

issue #5121 #7140

## What this PR does / why we need it: